[gui] mv main function inside init to avoid circular kind import when running python -m pairinteraction_gui
