+++ src/pairinteraction_gui/__init__.py
@@ -1,6 +1,28 @@
 # SPDX-FileCopyrightText: 2025 Pairinteraction Developers
 # SPDX-License-Identifier: LGPL-3.0-or-later
 
-from pairinteraction_gui.__main__ import main
+
+import sys
+
+from pairinteraction_gui.app import Application
+from pairinteraction_gui.main_window import MainWindow
 
 __all__ = ["main"]
+
+
+def main() -> int:
+    """Run the PairInteraction GUI application.
+
+    Returns:
+        int: Application exit code
+
+    """
+    app = Application(sys.argv)
+    app.setApplicationName("PairInteraction")
+
+    app.allow_ctrl_c()
+
+    window = MainWindow()
+    window.show()
+
+    return app.exec()
