+++ src/pairinteraction_gui/__main__.py
@@ -3,27 +3,7 @@
 
 import sys
 
-from pairinteraction_gui.app import Application
-from pairinteraction_gui.main_window import MainWindow
-
-
-def main() -> int:
-    """Run the PairInteraction GUI application.
-
-    Returns:
-        int: Application exit code
-
-    """
-    app = Application(sys.argv)
-    app.setApplicationName("PairInteraction")
-
-    app.allow_ctrl_c()
-
-    window = MainWindow()
-    window.show()
-
-    return sys.exit(app.exec())
-
+from pairinteraction_gui import main
 
 if __name__ == "__main__":
-    main()
+    sys.exit(main())
